+++ examples/date-picker/default.jsx
@@ -1,14 +1,25 @@
-import React from 'react';
+import React, { PropTypes } from 'react';
 import Datepicker from '~/components/date-picker';
 
 const Example = React.createClass({
 	displayName: 'DatepickerExample',
+	propTypes: {
+		log: PropTypes.func
+	},
 
 	render () {
 		return (
 			<Datepicker
-				onDateChange={function (date) {
-					console.log('>>> onDateChange ', date);
+				onChange={(event, data) => {
+					if (this.props.log) {
+						this.props.log('onChange')(
+						data.date,
+						data.formattedDate,
+						data.localOffset);
+					}
+				}}
+				onCalendarFocus={(event, data) => {
+					if (this.props.log) { this.props.log('onCalendarFocus')(event, data.date); }
 				}}
 			/>
 		);